imgproc/quadtree/restermask: ushort -> uint; dumping to cvMat with sub one sized pixel; faster invert
git-svn-id: https://scm.citationtech.net/trunk/shared/imgproc/imgproc@9308 b7e6147c-0f76-4c84-a09d-19069afaf10c

+++ imgproc/rastermask/cvmat.hpp
@@ -10,13 +10,13 @@ namespace imgproc {
 
 namespace bitfield {
 
-cv::Mat asCvMat(const RasterMask &mask, int pixelSize = 1);
+cv::Mat asCvMat(const RasterMask &mask, double pixelSize = 1);
 
 } // namespace bitfield
 
 namespace quadtree {
 
-cv::Mat asCvMat(const RasterMask &mask, int pixelSize = 1);
+cv::Mat asCvMat(const RasterMask &mask, double pixelSize = 1);
 
 } // namespace quadtree
